rename components & correct file structure& add main-header component
