+++ project/src/components/app/app.tsx
@@ -2,7 +2,7 @@ import Main from '../main/Main';
 
 function App(): JSX.Element {
   return (
-    <Main />
+    <Main genre="Drama" name="The Grand Budapest Hotel" year="2014" />
   );
 }
