--- project/src/components/types.ts
@@ -1,4 +1,0 @@
-export type CardData = {
-  title: string,
-  link: string,
-};
