+++ project/src/types/index.ts
@@ -1,0 +1,12 @@
+type Year = string | number;
+
+export type CardData = {
+  title: string,
+  link: string,
+};
+
+export type FilmInfoProps = {
+  genre: string,
+  name: string,
+  year: Year,
+}
